Reconfigure completor.vim
The auto popup has been historically nice for filename completion, since
I get a quick overview of the files in a directory when I type ../ -- I
can quickly check if I've already reached the proper directory if I need
to go higher.

However, the completeopt settings have to be set for auto popup to work
well. Specifically, noselect must be added at the very least, to make
sure the auto popup doesn't actually insert a candidate as it pops up.

The problem with that is that it makes repeated <C-x><C-p> and <C-l>
completion annoying; I have to <C-x><C-p>, select the entry, <C-x><C-p>,
select the entry, etc. instead of simply spamming
<C-x><C-p><C-x><C-p><C-x><C-p>.

So, I let completor.vim manage *only* filename completion (where it's
useful), disable it everywhere else, and attach an autocmd to make sure
my completeopt is reset when necessary.