Clean up plugins
There are a few plugins I notice I never :packadd anymore (FastFold,
sleuth) so I removed them.

There are also a few plugins in start that I don't rely on anymore
(YankStack, eunuch) so I removed them from the script as well, but
they're currently sitting in opt directories on my machine right now, in
case I miss their behavior.

I've mostly just gotten used to using named registers for stuff I want
to save (this has been going on for a while now) but YankStack config
was left in mostly out of habit.

vim-eunuch is a different case: I only ever installed it for the nicer
:SudoWrite, and I was mildly delighted about its automatic chmod upon
saving a file with a shebang. Both behaviors are convenient but can be
easily done without a plugin (:w ! sudo tee %, :! chmod +x %), plus
vim-eunuch comes with a lot of stuff I don't use, so I opted to remove
it.

Editorconfig has been unused so far, but it *is* the standard we use
here at work (I just haven't worked on any group projects recently). I
chose not to remove it, but place it in an opt directory so it doesn't
impact startup time.