Extend foldmethod mappings for frequent values
I'm most likely to use my <Leader>vd map to set foldmethod to syntax, so
I might as well make a mapping for that specific use case.

I also use markers to fold certain Vimscript, so a mapping for that is
also in order.

Finally, a mapping to set it back to manual, in case I accidentally
choose a foldmethod for the wrong buffer or something.