Add map for passing handler props
All too often I'm passing down a handler function as props in React.
Because of the style I follow, it's always in the following format:

onAction={handleAction}

I added a quick Insert-mode map for this. It's pretty hacky right now,
relying on a :s with a subsequent :let @/ = "" to clear search.

Better to go with a function moving forward.