Remove alpha from nrformats
So... adding alpha to nrformats removed a useful behavior for <C-a> that
I didn't know I was relying on: it automatically jumps to the next
number.

I'll likely just set it whenever I need to increment letters, which
isn't as frequently as I might think (just for creating ordered lists,
and very occasionally for writing out the alphabet).