Remove some of vim-qf's automatic behavior
vim-qf automatically opens the quickfix upon catching output, and also
automatically resizes the quickfix window to fit its contents.

I'm sure it fits well with other people's use case but it's too annoying
for me, so I disabled it.

Also, I added a convenience mapping for running AsyncMake on the current
file only, to make it easier to invoke a linter or similar.