Add quickfix unseen flag to statusline
With the help of some of vim-qf's helper functions, it's pretty easy to
toggle a g:quickfix_pending variable based on when there is unseen
output in the quickfix.

This is nice, because I can kick off an async lint and be notified when
there are errors, but I'm not taken out of my workflow (such as if the
quickfix auto-opened, which is why I disabled that behavior).

I simply add a statusline flag that appears only if g:quickfix_pending
is set. The flag mimics the modified flag, since that usually grabs my
eye well enough.

A loclist version would be nice, but it's not as simple as for QF (I'd
have to keep track of which loclist windows belong to which main windows).

Maybe that's a task for next time.