Use wildmode=list:full for first Tab
I preferred list:longest for the longest time because it "felt right" to
me that the first match not be preferred over the others.

However, now that I'm using vim-projectionist, it's causing unnecessary
extra Tabs after an :Ecomponent (otherwise, I'm left with a component
whose name is the longest common substring of the matches; not good!).

Setting it to list:full ought to deal with it nicely.