Set g:netrw_altfile
With g:netrw_altfile = 1, I can move up to a directory with vinegar's -,
navigate to a new file, open it, and have the alternate buffer be the
previous file I had open.

This means I can <C-^> between the two files regardless of how many
intermediary directory buffers I opened, as well as refer to the other
with #.

While I very rarely use netrw to browse to a file, on the rare
circumstances that I - up from a buffer (even accidentally!), I lose the
alternate buffer setup, which is aggravating because I've come to rely
on <C-^> for certain setups.