Move Scratch buffer behavior to plugin script
Since the Scratch buffer behavior is pretty self-contained, I'd like to
move it to a plugin script.

I'm messing around with dza's MRU mapping to make use of Scratch, so
having it in a separate plugin script rather than a snippet in my vimrc
ought to make it easier to establish dependency between them.