Add mru.vim
I revised an MRU mapping snippet from dza on #vim. I made it into a
command instead, that opens MRU in a split. Add a bang to open it in a
new tab, as it was originally.

I also make use of scratch.vim, instead of pass opts to tabnew as it was
originally.

The extra <CR> mapping to open a file was dropped; gf works just fine.