Disable pre-processor detection in vim-vue
The performance penalty for pre-processor detection in vim-vue is simply
too much; the lag in a Vue file is extremely annoying, especially since
my Vim is so snappy in all other cases.

So, I disabled this feature. It's not as nice because I don't have
proper syntax highlighting for Pug templates, but that's at least
bearable. Lag is not.