Add Visual line mode map for replaying macro
This is a neat idea I've seen a few times floating around other people's
configs, the latest of which is likely robertmeta's.

The implementation is basically just running normal! @ on the selection
range, which will run a macro on each line.

A bit of logic is done to guard for Visual line mode (won't make much
sense for Visual mode, might do unexpected things for Visual block
mode).