Tweak inchlsearch.vim
The <silent> arg to the maps for g/ and g? turned out to be unnecessary,
since any output from the :let command would have been overwritten by /
anyway. Furthermore, it caused g/ to not display the commandline until
after the first char was written.

This has been fixed.

Additionally, the augroup for inchlsearch has been named more aptly to
IncHlSearch.