Add split.vim
romainl on #vim has a gist for a universal opposite of J that splits
lines. split.vim is the aforementioned gist (with names replaced, since
it used to be called BreakHere).

I map this action to gS so as not to shadow any builtin keys, but at the
same time still be a mnemonic; gJ joins, gS splits.