Add CSniplets for functions in javascript.vim
vim-sandwich has a nice feature for surrounding with functions: gsiwf
will start a commandline and prompt for a function name, which will be
used to surround the word under the cursor.

Since the function name is entered through the commandline, CSniplets
apply. I use convenience CSniplets for JSON.stringify, console.warn, and
console.log since those are likely candidates for surrounding arbitrary
variables in Javascript buffers.