Move relative-path completion to javascript.vim
There used to be a snippet in my vimrc that was in charge of
automatically cd'ing to the directory containing the currently edited
buffer (and automatically restoring after leaving Insert mode), so that
I could use <C-x><C-f> completion after a relative path such as
../../<C-x><C-f>.

I recently realized this could go wrong when it started erroring out
with Gina buffers, since the gina://... buffer name makes Vim expand %:h
to gina:/ or similar, which is obviously not a directory and thus makes
the cd error out *every time I enter Insert mode*.

As a temporary solution, I moved the snippet to
after/ftplugin/javascript.vim so I know only Javascript buffers will
have this behavior. This is fine, since I only really use this behavior
for completing module paths for require/import in Javascript.