Revert some Git operations back to fugitive
Gina status -s was nice, but the diff from Gstatus (D) is a lot better
than the diff and compare functions in Gina (dd and !!), so I went back
to Gstatus.

Gina commit also inexplicably uses its own filetype, gina-commit,
instead of the perfectly good gitcommit for which niceties such as
formatoptions and textwidth are already set up properly.