Change SearchInPlace cgn map to gz
Previously, I had <Space>c be an operator to quickly cgn the next
motion. This was fine, but a g-prefixed mapping seemed a lot more
fitting for an operator than a <Space>-prefixed one.

gc was already taken by commentary, so I used gz instead (in my head, gz
sounds like an alternate version of gc if pronounced out loud).