Use strftime instead of system for date mappings
The external call to date makes my mappings noticeably slower than using
strftime, likely because of the added overhead of starting a new
program.

Re: portability, I'm not actually sure if strftime is better. Relying on
GNU date isn't portable for sure, so at least that's out of the way, but
by the sound of the docs strftime is affected by libs at compile time.
Might be a point to look into in the future.