Add loclist unseen flag to statusline
In the same vein as commit 390191d807c2cef291f1caf0ec0af0ede6591576,
I've added an [L] flag to my statusline whenever there are unseen errors
in the location list. This works window-local, as one would expect
loclist-based features to work.

This pairs extremely nicely with AsyncDo's loclist support (commit
cf9b0ea2e57c98e7a7d974a3909797f1798eb494), which I've already
incorporated into working with Makery.vim.

Unfortunately, the same simple algorithm as with the quickfix flag
cannot be applied to the loc list.

Quickfix is global, so *any* quickfix command from *any* window will be
able to set g:quickfix_pending, and *any* quickfix window will be able
to unset it.

For loclist, however, it's not as simple as simply setting a
window-local w:loclist_pending. There are in fact two windows that are
relevant to this issue: the window from which a loclist command was run
and its corresponding loclist window.

The way I did this was to have the statusline#loclist() function query
whether there is relevant loclist output every time it's run. In
contrast to g:quickfix_pending, I instead store g:loclist_seen state in
a dict, indexed by window number. It's simple enough to set the "seen"
state of a loclist (with the help of vim-qf), which statusline#loclist()
uses to decide whether or not to show the flag.