Set guicursor to always use block
A fan favorite feature (which is a default for GVim) is to have the
cursor appear as a block in normal mode and as a thin bar in insert
mode.

Block cursor works well in visualizing how exactly normal mode commands
will behave (e.g. c, x, d), but the way the cursor works in insert mode
(press a char -> print it under the cursor -> move cursor to the right)
can be visually depicted as a thin bar that goes *between* character
blocks.

This point of view is consistent with all the operations (I think) and
helps visually explain some insert mode "oddities" (e.g.  moving cursor
one char left when escaping out of insert mode, moving cursor one char
right with a).

While so, I still don't like the thin bar cursor. I have no problem
visually understanding why the block cursor moves the way it does; the
switching from block to thin bar and back is just visual noise for me.

Block cursor for *all* modes is definitely the setting for me.