Add cnoremap for clearing command args
I got this idea from a snippet somewhere from romainl. The idea is to
have a mapping to clear a command line, but leave the first
whitespace-separated word intact. Very useful for clearing args from a
command.

This *could* be improved by checking for : cmdtype explicitly (I don't
ever need anything like this when searching with /, for example).
Perhaps another time, if it actually bugs me.