Add convenience TWH command for Vimwiki diary
Nowadays, I exclusively use my Vimwiki diary as a logbook for my time
spent on different tasks. A report on total working hours (TWH) at the
end of the day is useful for me, but is tedious to manually calculate
with a calculator program.

A crude function ought to deal with it. I just extract out the times
from a diary entry (I use `07:00-08:00` formatting for all of my times),
have qalc calculate the durations in hours, then sum them all together
(keeping in mind that break times marked by BREAK are not included).