Tweak highlight.vim for C, shell, and netrw
Preprocessor directive colors are all over the place; they've been made
the same as Include, which is magenta (to align with other keywords in
languages like Javascript, where Include statements are with the import
keyword).

For shell, some non-important stuff such as flags and command subs were
linked to colored highlight groups, and so were fixed manually.
Parameter expansions are important to keep track of, so they're
highlighted in cyan.

A single fix is made for netrw, to color executable files green, similar
to my settings for ls.