Make StatusLineNC differ from StatusLine
Apparently, if StatusLineNC and StatusLine have the same highlight, Vim
will proceed to fill the statusline with ^. The only other check for
this case is if the fillchars setting for stl and stlnc differ.

Since I don't want stl and stlnc to display anything other than a space,
and I don't want the current window to look any different from
non-current windows, I've elected to change the guibg for StatusLineNC
by a trivial amount simply to meet the condition that StatusLineNC and
StatusLine differ, without compromising how my current and non-current
windows look.