Add vim-scriptease to opt
I've found that time and again I'm looking for some functionality of
vim-scriptease. It's not all the time; I rarely configure Vim scripts
nowadays, relative to the amount of time I do actual work. But when I
*do* configure them, some of vim-scriptease's features would have been
nice.

But like I said, it's not all the time, so opt is the perfect place for
it.