Add mapping for searching same-indent blocks
When working with Pug files, I frequently find myself wanting to move to
siblings of an element, which translates to lines with the same level of
indentation in Pug.

I used to do different hacks to pull this off:

- cursorcolumn (works only if the two siblings appear one screen height
  apart at most)
- indent guides (same issue)
- yanking the sibling, finding a candidate, pasting the sibling next to
  it to see if they're the same level of indentation (uhh...)

I realized I could just map a motion for this and be done with it. I
pull this off by yanking the current line's indentation and simply
starting a search for it. I include ^ at the beginning of the match and
a non-whitespace char at the end to make sure I match only the exact
level of indentation, not a subset of a more indented line.

/pattern/e works very nicely here: it positions the cursor at the end of
a match instead of the beginning, so the cursor goes to the first
non-whitespace char of the line. Very handy!

It's funny how long I've coped with various hacks when there was a
simple solution for this all along: I just had to think like a Vimmer.