Reduce colors for HTML, Pug
Tag names are no longer colored magenta. This is actually useful
highlighting for HTML, but is unfortunately the same highlighting group
used in Pug, where it's just visual noise.

I've gotten very used to syntax off for Pug (since I've had Vue
preprocessor syntax highlighting off for so long), so much so that I
find tag name highlighting to no longer be readable in the same, flowing
(for lack of a better word) manner as with no highlighting at all.