Force nojump with asyncdo
How could I have not seen this for so long?!

Apparently asyncdo has a nojump argument that allows me to prevent
asyncdo#run from jumping to the first error after grabbing the result.

This behavior is just incredibly annoying, since I want to be able to do
other stuff while running something in the background; it's why I put it
in the background in the first place!

This, paired with my [Q] and [L] flags that notify me of new
quickfix/loclist output, makes for really smooth grepping, linting, and
overall task running.