Use Normal instead of Identifier as base highlight
For some reason, I linked all unhighlighted groups to Identifier
(probably because it was the first white text group that came to my
mind).

It makes more sense to link to Normal, since that's what I actually want
to do: bring these highlight groups in line with my normal
white-foreground-on-dark-background text.