Add highlights for i3 config
I noticed that a few highlight groups for i3 config
(i3ConfigVariableModifier and i3ConfigString specifically) had not been
configured to be in line with my colorscheme in general --
i3ConfigVariableModifier appeared cyan which leads to breaks in my
reading, while i3ConfigString appears my normal white on dark instead of
green (as is usual for my strings).

A few tweaks are in order.