Restructure cfg to represent home dir
In preparation for a few new Neovim-specific additions to my config, I'm
moving the cfg one directory up, to the home dir. This means all Vim
config now lives in cfg/.vim instead of immediately in cfg.

This allows me to create cfg/.config/nvim, where Neovim-specific config
can live. I'd prefer this for larger features such as LSP so as not to
litter my Vim files with if has('nvim') everywhere.